app-engine/standard/python37/private-vpc-access/requirements.txt
flask
redis
google-cloud-redis
google-api-python-client
